services/pdf-png/requirements.txt
pip>=25.2
setuptools>=75.1.0
wheel>=0.43
packaging>=24.1

Flask>=3.0.0
gunicorn>=21.2.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv==1.0.0

Pillow>=10.2.0
PyMuPDF>=1.24.9